+++ translator/src/main/scala/soda/lib/Recursion.scala
@@ -10,8 +10,8 @@ package soda.lib
 trait Recursion {
 
   import scala.annotation.tailrec
-        @tailrec  final
-  def _tailrec_fold4 [A, B, C <: B] (sequence: Seq [A], current_value: C, next_value_function: (B, A ) => C, condition: (B, A ) => Boolean         ): C =
+  @tailrec  final
+  def _tailrec_fold4 [A, B, C <: B] (sequence: Seq [A], current_value: C, next_value_function: (B, A ) => C, condition: (B, A ) => Boolean ): C =
     if (sequence.isEmpty
     ) current_value
     else
@@ -19,21 +19,21 @@ trait Recursion {
       ) current_value
       else _tailrec_fold4 (sequence.tail, next_value_function (current_value, sequence.head ), next_value_function, condition )
 
-  def fold [A, B, C <: B] (sequence: Seq [A], initial_value: C, next_value_function: (B, A ) => C, condition: (B, A ) => Boolean  ): C =
+  def fold [A, B, C <: B] (sequence: Seq [A], initial_value: C, next_value_function: (B, A ) => C, condition: (B, A ) => Boolean ): C =
     _tailrec_fold4 (sequence, initial_value, next_value_function, condition )
 
   import scala.annotation.tailrec
-        @tailrec  final
-  def _tailrec_fold3 [A, B, C <: B] (sequence: Seq [A], current_value: C, next_value_function: (B, A ) => C         ): C =
+  @tailrec  final
+  def _tailrec_fold3 [A, B, C <: B] (sequence: Seq [A], current_value: C, next_value_function: (B, A ) => C ): C =
     if (sequence.isEmpty
     ) current_value
     else _tailrec_fold3 (sequence.tail, next_value_function (current_value, sequence.head ), next_value_function )
 
-  def fold [A, B, C <: B] (sequence: Seq [A], initial_value: C, next_value_function: (B, A ) => C  ): C =
+  def fold [A, B, C <: B] (sequence: Seq [A], initial_value: C, next_value_function: (B, A ) => C ): C =
     _tailrec_fold3 (sequence, initial_value, next_value_function )
 
   import scala.annotation.tailrec
-        @tailrec  final
+  @tailrec  final
   def _tailrec_range (n: Int, sequence: Seq [Int]  ): Seq [Int] =
     if (n <= 0
     ) sequence