+++ translator/src/main/scala/soda/translator/parser/annotation/BlockAnnotation.scala
@@ -14,9 +14,6 @@ trait BlockAnnotation {
 
   lazy val space = SodaConstant_ () .space
 
-  lazy val contains_one_line: Boolean =
-    block.readable_lines.length == 1
-
   def starts_with_prefix_and_space (prefix: String ): Boolean =
     block.readable_lines.nonEmpty &&
     block.readable_lines.head.line.trim.startsWith (prefix + space )