+++ translator/src/main/scala/soda/translator/parser/annotation/ClassBeginningAnnotation.scala
@@ -8,7 +8,6 @@ trait ClassBeginningAnnotation  extends BlockAnnotation {
   lazy val identifier = BlockAnnotationEnum_ () .class_beginning
 
   lazy val applies: Boolean =
-    contains_one_line &&
     (starts_with_prefix_and_space (SodaConstant_ () .class_reserved_word ) ||
       starts_with_prefix_and_space (SodaConstant_ () .class_abbreviation ) ) &&
     ends_with_space_and_suffix (SodaConstant_ () .class_open_symbol )