+++ translator/src/main/scala/soda/translator/parser/annotation/ClassDeclarationAnnotation.scala
@@ -8,7 +8,6 @@ trait ClassDeclarationAnnotation  extends BlockAnnotation {
   lazy val identifier = BlockAnnotationEnum_ () .class_declaration
 
   lazy val applies: Boolean =
-    contains_one_line &&
     (starts_with_prefix_and_space (SodaConstant_ () .class_reserved_word ) ||
       starts_with_prefix_and_space (SodaConstant_ () .class_abbreviation ) ) &&
     ! ends_with_space_and_suffix (SodaConstant_ () .class_open_symbol )