+++ translator/src/main/scala/soda/translator/parser/annotation/ClassEndAnnotation.scala
@@ -8,7 +8,6 @@ trait ClassEndAnnotation  extends BlockAnnotation {
   lazy val identifier = BlockAnnotationEnum_ () .class_end
 
   lazy val applies: Boolean =
-    contains_one_line &&
     block.readable_lines.nonEmpty &&
     (block.readable_lines.head.line.trim == SodaConstant_ () .class_close_symbol )