+++ translator/src/main/scala/soda/translator/parser/annotation/FunctionDefinitionAnnotation.scala
@@ -2,10 +2,10 @@ package soda.translator.parser.annotation
 
 trait FunctionDefinitionAnnotation  extends BlockAnnotation {
 
-  import soda.translator.block.Block
-  import soda.translator.block.Block_
   import soda.translator.block.BlockAnnotationEnum_
   import soda.translator.parser.SodaConstant_
+  import soda.translator.replacement.ParserStateEnum_
+  import soda.translator.replacement.Tokenizer_
 
   lazy val identifier = BlockAnnotationEnum_ () .function_definition
 
@@ -29,18 +29,27 @@ trait FunctionDefinitionAnnotation  extends BlockAnnotation {
     sc.function_definition_synonym +
     sc.space
 
+  lazy val plain_state = ParserStateEnum_ () .plain
+
   lazy val applies: Boolean =
     (contains_the_equals_symbol || starts_with_valid_annotation ) && ! is_a_class_declaration
 
   lazy val contains_the_equals_symbol: Boolean =
     block.readable_lines.nonEmpty &&
-    _contains_the_equals_symbol_with (block.readable_lines.head.line.trim )
+    _contains_the_equals_symbol_with (block.readable_lines.head.line )
 
-  def _contains_the_equals_symbol_with (first_line_trimmed: String ): Boolean =
-    ((contains_one_line && first_line_trimmed.contains (symbol_in_the_middle ) ) ||
-      (contains_one_line && first_line_trimmed.contains (synonym_in_the_middle ) ) ||
-      (first_line_trimmed.endsWith (symbol_at_the_end ) ) ||
-      (first_line_trimmed.endsWith (synonym_at_the_end ) )    )
+  def _contains_the_equals_symbol_with (first_line: String ): Boolean =
+    Tokenizer_ (first_line )
+      .tokens
+      .exists (token =>
+        token.parser_state == plain_state &&
+        _contains_the_equals_symbol_in_token (token.text )      )
+
+  def _contains_the_equals_symbol_in_token (token_text: String ): Boolean =
+    ((token_text.contains (symbol_in_the_middle ) ) ||
+      (token_text.contains (synonym_in_the_middle ) ) ||
+      (token_text.endsWith (symbol_at_the_end ) ) ||
+      (token_text.endsWith (synonym_at_the_end ) )    )
 
   lazy val starts_with_valid_annotation: Boolean =
     block.readable_lines.nonEmpty &&