+++ translator/src/main/scala/soda/translator/parser/annotation/PackageDeclarationAnnotation.scala
@@ -8,7 +8,6 @@ trait PackageDeclarationAnnotation  extends BlockAnnotation {
   lazy val identifier = BlockAnnotationEnum_ () .package_declaration
 
   lazy val applies: Boolean =
-    contains_one_line &&
     starts_with_prefix_and_space (SodaConstant_ () .package_reserved_word )
 
 }