+++ docs/RELEASE-NOTES.md
@@ -1,6 +1,9 @@
 ---
 - - version: v0.14.0
   - date: unreleased
+  - new_features:
+    - produces a more similar translation to Scala by removing the line joiners
+    - restricts how a constant or function is defined, such that the definiendum cannot be written in multiple lines
   - build: sbt '++ 3.1.0' clean compile test package assembly
   - release: target/scala-3.1.0/soda-0.14.0.jar
 - - version: v0.13.0